+++ example/android/app/src/main/kotlin/com/whisper_ggml/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.whisper_ggml.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()