+++ example/linux/flutter/generated_plugins.cmake
@@ -3,11 +3,13 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  ffmpeg_kit_flutter_new_min
   record_linux
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
-  whisper_ggml
+  jni
+  whisper_wrapper
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)