+++ lib/src/models/responses/whisper_transcribe_segment.g.dart
@@ -7,21 +7,23 @@ part of 'whisper_transcribe_segment.dart';
 // **************************************************************************
 
 _WhisperTranscribeSegment _$WhisperTranscribeSegmentFromJson(
-        Map<String, dynamic> json) =>
-    _WhisperTranscribeSegment(
-      fromTs: WhisperTranscribeSegment._durationFromInt(
-          (json['from_ts'] as num).toInt()),
-      toTs: WhisperTranscribeSegment._durationFromInt(
-          (json['to_ts'] as num).toInt()),
-      text: json['text'] as String,
-      speakerTurnNext: json['speaker_turn_next'] as bool? ?? false,
-    );
+  Map<String, dynamic> json,
+) => _WhisperTranscribeSegment(
+  fromTs: WhisperTranscribeSegment._durationFromInt(
+    (json['from_ts'] as num).toInt(),
+  ),
+  toTs: WhisperTranscribeSegment._durationFromInt(
+    (json['to_ts'] as num).toInt(),
+  ),
+  text: json['text'] as String,
+  speakerTurnNext: json['speaker_turn_next'] as bool? ?? false,
+);
 
 Map<String, dynamic> _$WhisperTranscribeSegmentToJson(
-        _WhisperTranscribeSegment instance) =>
-    <String, dynamic>{
-      'from_ts': instance.fromTs.inMicroseconds,
-      'to_ts': instance.toTs.inMicroseconds,
-      'text': instance.text,
-      'speaker_turn_next': instance.speakerTurnNext,
-    };
+  _WhisperTranscribeSegment instance,
+) => <String, dynamic>{
+  'from_ts': instance.fromTs.inMicroseconds,
+  'to_ts': instance.toTs.inMicroseconds,
+  'text': instance.text,
+  'speaker_turn_next': instance.speakerTurnNext,
+};